pyproject.toml
[project]
name = "cleanup-helper-app"
version = "0.1.0"
description = ""
authors = [
    {name = "yoiyoi"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit (>=1.45.0,<2.0.0)",
    "opencv-python-headless (>=4.11.0.86,<5.0.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "ultralytics (>=8.3.123,<9.0.0)",
    "pytest (>=8.3.5,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
